Rename products serializers, views and urls, delete empty views file
