+++ backend/good_food/urls.py
@@ -5,7 +5,7 @@ from django.urls import include, path
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("api.urls_products")),
+    path("api/", include("api.products_urls")),
     path("api/", include("api.users_urls")),
 ]
